+++ src/index.js
@@ -1,7 +1,7 @@
 'use strict'
 
+const util = require('util')
 const fs = require('fs')
-const pify = require('pify')
 const locate = require('./locate')
 const ejs = require('./ejs')
 const data = require('./data')
@@ -20,7 +20,7 @@ module.exports = async function(filePath, opts) {
 
 	const dataPath = await locate(filePath)
 	const json = await data(dataPath, opts)
-	const str = await pify(fs.readFile)(filePath, 'utf8')
+	const str = await util.promisify(fs.readFile)(filePath, 'utf8')
 
 	return ejs(filePath, str, json)