+++ CHANGELOG.md
@@ -42,7 +42,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [8.0.0] - 2017-07-19
 
-### New
+### Added
 
 - Only support Node.js 7 and 8